🐛 fix : 특정 UI 에서 의도치 않은 재로그인 요청 해결 #307
